Programming Basic First Project - 4 Task
Program 1

Write a program that prints the numbers from 1 to 100. But for multiples
of three print “Foo” instead of the number and for the multiples of five
print “Bar”. For numbers which are multiples of both three and five
print “FooBar”.

Program 2

Write a program that will read a natural number from 1 to 7 and print
out a corresponding day o the week (1 -> Monday, 2 -> Tuesday, ..., 7 ->
Sunday). The number should be given by the user. Handle a situation when
the number is not in the preferred range.

Program 3

Write a program with three versions of a function int sum(int min, int
max) that will calculate a sum from a range:

of all the numbers,
only for even numbers,
only for numbers divisible by 3.
The main function should contain test invocation of all three functions.

Program 4

Write a program with a set of functions float avg_X(int n, int m, float
t1[n][m]) (where X is the number of the function) that will calculate an
average value of the t1 array elements for:

the entire array,
only the main diagonal,
a given column (the column number passed as an additional function
parameter).